backend/requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask_cors
psycopg2-binary
redis
flask-bcrypt
flask-jwt-extended